requirements.txt
streamlit
transformers
sentence-transformers
diffusers
huggingface_hub
torch
Pillow
moviepy
python-dotenv
numpy
